requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiohttp==3.9.1
schedule==1.2.1 